requirements.txt
accelerate==0.26.1
certifi==2023.11.17
charset-normalizer==3.3.2
diffusers==0.14.0
einops==0.7.0
filelock==3.13.1
fsspec==2023.12.2
huggingface-hub==0.20.2
idna==3.6
importlib-metadata==7.0.1
numpy==1.24.4
scipy==1.10.1
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
protobuf==4.25.2
psutil==5.9.7
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
tensorboardX==2.6.2.2
tokenizers==0.15.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.9.0
urllib3==2.1.0
zipp==3.17.0
